Implemented a new method of randomising songs. All songs IDs are retrieved at once and shuffled into an array. As more random songs are requested, song details are retrieved from the shuffled indices. This guarantees that each song in the library will be played once, until the array is re-shuffled. I should implement a class allowing different randomisation methods to be chosen and link it to user preference.
Added a method to check whether any songs exist in the Android library.
The status is cached because the query is slow on my phone. I think I
should implement a new, faster query using the SQL operator EXISTS
instead.

Class is notified when any media is added to the android library. Songs
are marked to be reshuffled and cached data is invalidated in response.
Reshuffling is probably too expensive an operation to do at this point
-- that should be down to user preference.